pyproject.toml
[tool.poetry]
name = "crypto-analytics"
version = "0.1.0"
description = "Análise de criptomoedas"
authors = ["matheusmafraoandrade <matheusmafrabh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
python-binance = "^1.0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
